requirements.txt
python-binance


